Provide two final versions of EditableStyledDocument
