+++ richtextfx/src/main/java/org/fxmisc/richtext/model/SimpleEditableStyledDocument.java
@@ -1,6 +1,6 @@
 package org.fxmisc.richtext.model;
 
-public final class SimpleEditableStyledDocument<PS, S> extends GenericEditableStyledDocument<PS, StyledText<S>, S> {
+public final class SimpleEditableStyledDocument<PS, S> extends GenericEditableStyledDocumentBase<PS, StyledText<S>, S> {
 
     public SimpleEditableStyledDocument(PS initialParagraphStyle, S initialStyle) {
         super(initialParagraphStyle, initialStyle, StyledText.textOps());
